refactoring step 1
