+++ application/linkkivinkit/src/main/java/linkkivinkki/app/command/Command.java
@@ -1,0 +1,6 @@
+
+package linkkivinkki.app.command;
+
+public interface Command {
+    boolean execute();
+}
