+++ application/linkkivinkit/src/main/java/linkkivinkki/app/command/Invalid.java
@@ -1,0 +1,20 @@
+
+package linkkivinkki.app.command;
+
+import linkkivinkki.io.Color;
+import linkkivinkki.io.IO;
+
+public class Invalid implements Command {
+    private IO io;
+
+    public Invalid(IO io) {
+        this.io = io;
+    }
+
+    @Override
+    public boolean execute() {
+        io.print(Color.redText("Invalid command."));
+        return false;
+    }
+
+}
